+++ infra/postgres/migrations/0043_workflow_template.down.sql
@@ -1,0 +1,7 @@
+BEGIN;
+
+DROP TABLE IF EXISTS workflow_template_use;
+DROP TABLE IF EXISTS workflow_template;
+DROP TABLE IF EXISTS workflow_template_occurrence;
+
+COMMIT;